--- backend/cms-sys/.env
@@ -1,21 +1,0 @@
-# Application
-APP_ENV=development
-PORT=8080
-
-# Logging
-LOG_LEVEL=info
-LOG_FILE_PATH=logs/app.log
-LOG_MAX_SIZE=100
-LOG_MAX_BACKUPS=5
-LOG_MAX_AGE=30
-LOG_COMPRESS=true
-LOG_CONSOLE=true
-
-# Database
-DB_HOST=postgres-cms
-DB_PORT=5432
-DB_USER=cms_user
-DB_PASSWORD=cms_password
-DB_NAME=cms_db
-DB_SSL_MODE=disable
-
