+++ backend/cms-sys/.env.cms
@@ -1,0 +1,49 @@
+# Application
+CMS_APP_STATUS=production
+CMS_PORT=8081
+CMS_AllOWED_ORIGIN=http://localhost:3000
+
+# Logger
+CMS_LOG_LEVEL=info
+CMS_LOG_FILE_PATH=logs/app.log
+CMS_LOG_MAX_SIZE=100
+CMS_LOG_MAX_BACKUPS=5
+CMS_LOG_MAX_AGE=30
+CMS_LOG_COMPRESS=true
+CMS_LOG_CONSOLE=true
+
+# JWT
+JWT_SECRET=your_jwt_secret
+
+# AWS & SSM (for production mode)
+AWS_REGION=us-east-1
+AWS_ACCESS_KEY_ID=your_access_key_id
+AWS_SECRET_ACCESS_KEY=your_secret_key
+
+# Consul Configuration
+CONSUL_ENABLED=true
+CMS_CONSUL_ADDRESS=consul:8500
+CMS_CONSUL_DATACENTER=dc1
+CMS_CONSUL_TOKEN=
+CMS_CONSUL_SCHEME=http
+CMS_CONSUL_SERVICE_ID=cms-doc-api-localhost
+CMS_CONSUL_SERVICE_NAME=cms-service
+CONSUL_SERVICE_TAGS=cms,multi-tenant,api
+CONSUL_CHECK_TTL=30s
+
+# Database
+CMS_DB_HOST=localhost
+CMS_DB_PORT=5432
+CMS_DB_USER=postgres
+CMS_DB_PASSWORD=Swanhtet12@
+CMS_DB_NAME=cms_db
+CMS_DB_SSL_MODE=disable
+CMS_DB_MAX_OPEN_CONNS=25
+CMS_DB_MAX_IDLE_CONNS=10
+CMS_DB_CONN_MAX_LIFETIME=5m
+CMS_DB_CONN_MAX_IDLE_TIME=2m
+CMS_DB_RETRY_ATTEMPTS=5
+CMS_DB_RETRY_DELAY=2s
+
+# S3
+BUCKET_NAME=your_bucket_name
