--- backend/file/file-upload-s3/go.mod
@@ -1,3 +1,0 @@
-module github.com/multi-tenants-cms-golang/file-upload-s3
-
-go 1.24
