--- backend/file/file-upload-s3/main.go
@@ -1,5 +1,0 @@
-package main
-
-func main() {
-
-}
