+++ backend/lms-main-system/app/cornServer/handler/databasebackHandler.go
@@ -6,26 +6,12 @@ import (
 	"fmt"
 	"github.com/hibiken/asynq"
 	"github.com/jackc/pgx/v5"
-	"github.com/klauspost/compress/gzip"
 	"github.com/multi-tenants-cms-golang/lms-sys/pkg/task"
 	"github.com/sirupsen/logrus"
-	"io"
 	"os"
 	"path/filepath"
 	"time"
 )
-
-type CornHandler struct {
-	logger *logrus.Logger
-}
-
-func NewCornHandler(
-	logger *logrus.Logger,
-) *CornHandler {
-	return &CornHandler{
-		logger: logger,
-	}
-}
 
 func (handler *CornHandler) HandleDatabaseBackUp(ctx context.Context, req *asynq.Task) error {
 	var payload task.DatabaseBackup
@@ -139,48 +125,5 @@ func (handler *CornHandler) HandleDatabaseBackUp(ctx context.Context, req *asynq
 	}
 
 	handler.logger.Infof("PostgreSQL backup completed successfully: %s", backupFileName)
-	return nil
-}
-
-func compressBackupFile(filename string) error {
-	original, err := os.Open(filename)
-	if err != nil {
-		return err
-	}
-	defer func(original *os.File) {
-		err := original.Close()
-		if err != nil {
-			logrus.Errorf("Failed to close gzip writer: %v", err)
-		}
-	}(original)
-
-	compressed, err := os.Create(filename + ".gz")
-	if err != nil {
-		return err
-	}
-	defer func(compressed *os.File) {
-		err := compressed.Close()
-		if err != nil {
-			logrus.Errorf("Failed to close gzip writer: %v", err)
-		}
-	}(compressed)
-
-	gz := gzip.NewWriter(compressed)
-	defer func(gz *gzip.Writer) {
-		err := gz.Close()
-		if err != nil {
-			logrus.Errorf("Failed to close gzip writer: %v", err)
-		}
-	}(gz)
-
-	if _, err := io.Copy(gz, original); err != nil {
-		return err
-	}
-
-	err = os.Remove(filename)
-	if err != nil {
-		logrus.Errorf("Failed to remove backups file: %v", err)
-		return err
-	}
 	return nil
 }
