+++ backend/lms-main-system/app/rpc/authentication/rpc_email_resend.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-func (s *AuthenticationService) ResendVerificationEmail(
+func (s *Service) ResendVerificationEmail(
 	ctx context.Context,
 	req *authenticationpb.ResendVerificationRequest,
 ) (*authenticationpb.ResendVerificationResponse, error) {
