+++ backend/lms-main-system/cmd/main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"context"
-	"fmt"
-	"github.com/jackc/pgx/v5"
-	"github.com/multi-tenants-cms-golang/lms-sys/app/cornServer"
-	"github.com/multi-tenants-cms-golang/lms-sys/app/cornServer/handler"
-	db "github.com/multi-tenants-cms-golang/lms-sys/internal/repo"
-	"github.com/multi-tenants-cms-golang/lms-sys/pkg/utils/nats"
-	"github.com/multi-tenants-cms-golang/lms-sys/pkg/utils/redis"
+	"github.com/joho/godotenv"
+	"github.com/multi-tenants-cms-golang/lms-sys/pkg/infra/consul"
+	"github.com/natefinch/lumberjack"
+	"github.com/sirupsen/logrus"
 	"io"
 	"os"
 	"runtime"
@@ -17,38 +13,61 @@ import (
 
 	"github.com/hashicorp/consul/api"
 	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/joho/godotenv"
+
 	"github.com/multi-tenants-cms-golang/lms-sys/app"
+	"github.com/multi-tenants-cms-golang/lms-sys/app/cornServer"
+	"github.com/multi-tenants-cms-golang/lms-sys/app/cornServer/handler"
 	"github.com/multi-tenants-cms-golang/lms-sys/app/gateway"
 	"github.com/multi-tenants-cms-golang/lms-sys/app/rpc"
+	db "github.com/multi-tenants-cms-golang/lms-sys/internal/repo"
+	"github.com/multi-tenants-cms-golang/lms-sys/pkg/infra/nats"
+	"github.com/multi-tenants-cms-golang/lms-sys/pkg/infra/postgres"
+	"github.com/multi-tenants-cms-golang/lms-sys/pkg/infra/redis"
 	"github.com/multi-tenants-cms-golang/lms-sys/pkg/utils/env"
-	"github.com/natefinch/lumberjack"
-	"github.com/sirupsen/logrus"
 )
 
-func initLogger() *logrus.Logger {
-	if err := godotenv.Load(); err != nil {
-		logrus.WithError(err).Fatal("Failed to load .env")
-	}
+func main() {
+	logger := initLogger()
+	defer closeLogger(logger)
 
+	loadEnv(logger)
+
+	dbPool := initDatabase(logger)
+	defer dbPool.Close()
+
+	initRedis(logger)
+	initNATS(logger)
+
+	serviceID := env.GetEnv("LMS_SERVICE_ID", "lms-srvice-1")
+	consulClient := initConsul(logger, serviceID)
+
+	app := initApp(logger, dbPool, consulClient)
+
+	if err := app.Run(); err != nil {
+		logger.WithError(err).Fatal("Failed to start server")
+	}
+}
+
+func loadEnv(logger *logrus.Logger) {
+	if err := godotenv.Load(); err != nil {
+		logger.WithError(err).Fatal("Failed to load .env file")
+	}
+	logger.Info("Environment variables loaded")
+}
+
+func initLogger() *logrus.Logger {
 	logger := logrus.New()
-	logger.Info("Initializing logger with rotation and JSON formatting")
 
 	logFile := &lumberjack.Logger{
 		Filename:   "logs/lms-system.log",
 		MaxSize:    100,
 		MaxBackups: 3,
-		MaxAge:     28, // days
+		MaxAge:     28,
 		Compress:   true,
 		LocalTime:  true,
 	}
-	logger.Infof("Configured log rotation: file=%s, maxSize=%dMB, maxBackups=%d, maxAge=%ddays, compress=%t",
-		logFile.Filename, logFile.MaxSize, logFile.MaxBackups, logFile.MaxAge, logFile.Compress)
-
-	logger.Info("Setting dual output to stdout and log file")
 	logger.SetOutput(io.MultiWriter(os.Stdout, logFile))
 
-	logger.Info("Configuring JSON formatter with custom fields")
 	logger.SetFormatter(&logrus.TextFormatter{
 		TimestampFormat: "2006-01-02T15:04:05.999Z07:00",
 		FieldMap: logrus.FieldMap{
@@ -58,238 +77,89 @@ func initLogger() *logrus.Logger {
 			logrus.FieldKeyFunc:  "caller",
 		},
 		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
-			file := strings.TrimPrefix(f.File, "/Users/swanhtet/Desktop/multi-tenants-cms-golang/backend/lms-main-system/")
 			funcName := f.Function[strings.LastIndex(f.Function, ".")+1:]
-			return file, funcName + ":" + string(rune(f.Line))
+			return f.File, funcName + ":" + string(rune(f.Line))
 		},
 		ForceColors: true,
 	})
-
-	// Enable caller reporting
-	logger.Info("Enabling caller reporting in logs")
 	logger.SetReportCaller(true)
-
-	// Set log level from environment
-	logLevel := os.Getenv("LOG_LEVEL")
-	if logLevel == "" {
-		logLevel = "info"
-	}
-	logger.Infof("Setting log level: %s", logLevel)
-
-	level, err := logrus.ParseLevel(logLevel)
-	if err != nil {
-		logger.Warnf("Invalid LOG_LEVEL '%s', defaulting to 'info'", logLevel)
-		level = logrus.InfoLevel
-	}
-	logger.SetLevel(level)
-
-	logger.WithFields(logrus.Fields{
-		"logLevel":    logger.GetLevel(),
-		"outputs":     "stdout+file",
-		"formatter":   "JSON",
-		"callerInfo":  true,
-		"compression": logFile.Compress,
-	}).Info("Logger initialization complete")
-
 	return logger
 }
-func main() {
-	logger := initLogger()
-	defer func() {
-		if file, ok := logger.Out.(*lumberjack.Logger); ok {
-			err := file.Close()
-			if err != nil {
-				return
-			}
-		}
-	}()
 
-	if err := godotenv.Load(); err != nil {
-		logger.WithError(err).Fatal("Failed to load .env")
+func closeLogger(logger *logrus.Logger) {
+	if file, ok := logger.Out.(*lumberjack.Logger); ok {
+		_ = file.Close()
 	}
-	logger.Info("Application starting with log rotation enabled")
+}
 
-	grpcServerAddress := env.GetEnv("LMS_GRPC_SERVER_ADDRESS", ":9001")
-	//grpcGatewayAddress := env.GetEnv("LMS_GRPC_GATEWAY_ADDRESS", ":8086")
-	jwtSecret := env.GetEnv("LMS_JWT_SECRET", "default-secret-key-at-least-32-characters-long")
-	jwtIssuer := env.GetEnv("LMS_JWT_ISSUER", "lms-system")
-	jwtAudience := env.GetEnv("LMS_JWT_AUDIENCE", "lms-client")
-	consulAddress := env.GetEnv("LMS_CONSUL_ADDRESS", "localhost:8500")
-	redisAddress := env.GetEnv("LMS_REDIS_ADDRESS", "localhost:6379")
-	redisPassword := env.GetEnv("LMS_REDIS_PASSWORD", "")
-	dbRedis := env.GetEnvAsInt("LMS_DB_REDIS_ADDRESS", 0)
+func initDatabase(logger *logrus.Logger) *pgxpool.Pool {
+	connectionPool := postgres.DatabaseConn(logger)
+	logger.Info("Database connection initialized")
+	return connectionPool
+}
+
+func initRedis(logger *logrus.Logger) {
+	redisAddr := env.GetEnv("LMS_REDIS_ADDRESS", "localhost:6379")
+	redisPass := env.GetEnv("LMS_REDIS_PASSWORD", "")
+	dbNum := env.GetEnvAsInt("LMS_DB_REDIS_ADDRESS", 0)
+
+	redis.InitRedis(redisAddr, redisPass, dbNum)
+	logger.Info("Redis initialized")
+}
+
+func initNATS(logger *logrus.Logger) {
 	natsUrl := env.GetEnv("LMS_NATS_URL", "nats://localhost:4222")
-	//serviceName := env.GetEnv("LMS_SERVICE_NAME", "lms-service")
-	serviceID := env.GetEnv("LMS_SERVICE_ID", "lms-srvice-1")
-
-	dbConn := DatabaseConn(logger)
-	defer dbConn.Close()
-
-	redis.InitRedis(redisAddress, redisPassword, dbRedis)
-	err := nats.InitNATS(natsUrl)
-	if err != nil {
+	if err := nats.InitNATS(natsUrl); err != nil {
 		logger.WithError(err).Fatal("Failed to connect to NATS")
-		return
 	}
-	dbUrl := "postgresql://neondb_owner:npg_MEB4CYJS7TKh@ep-withered-salad-a27oo7gn-pooler.eu-central-1.aws.neon.tech/neondb?sslmode=require&channel_binding=require"
+	logger.Info("NATS initialized")
+}
 
-	backupHanlder := handler.NewCornHandler(logger)
-	cronServer := cornServer.NewServer(
-		logger,
-		redisAddress,
-		backupHanlder,
-		dbUrl,
-	)
-	dbStore := db.NewStore(logger, dbConn)
-
-	consulClient, err := NewConsulClient(consulAddress)
+func initConsul(logger *logrus.Logger, serviceID string) *api.Client {
+	consulAddr := env.GetEnv("LMS_CONSUL_ADDRESS", "localhost:8500")
+	client, err := consul.NewConsulClient(consulAddr)
 	if err != nil {
 		logger.WithError(err).Fatal("Failed to create Consul client")
 	}
-	defer func(client *api.Client, serviceID string) {
-		err := DeregisterService(client, serviceID)
-		if err != nil {
-			logger.WithError(err).Fatal("Failed to deregister service")
-		}
-	}(consulClient, serviceID)
+	logger.Info("Consul client initialized")
 
-	//if err := RegisterService(consulClient, serviceName, serviceID+"-grpc", grpcServerAddress, "grpc"); err != nil {
-	//	logger.WithError(err).Fatal("Failed to register gRPC server with Consul")
-	//}
-	//
-	//if err := RegisterService(consulClient, serviceName, serviceID+"-gateway", grpcGatewayAddress, "http"); err != nil {
-	//	logger.WithError(err).Fatal("Failed to register gRPC gateway with Consul")
-	//}
+	serviceName := env.GetEnv("LMS_SERVICE_NAME", "lms-service")
+	serviceTags := env.GetEnv("LMS_SERVICE_TAGS", "lms,learning,api,v1,environment:development")
+	serviceTagsSlice := strings.Split(serviceTags, ",")
+	serviceAddress := env.GetEnv("LMS_SERVICE_ADDRESS", "localhost:8086")
+	serviceType := "http"
 
-	grpcServer := rpc.NewServer(
-		dbStore,
-		logger,
-		jwtSecret,
-		jwtIssuer,
-		jwtAudience,
-		grpcServerAddress,
-	)
-
-	grpcGateway := gateway.NewGateway(
-		logger,
-		grpcServerAddress,
-		":8086",
-		//jwtSecret,
-		//jwtIssuer,
-		//jwtAudience,
-	)
-
-	server := app.NewApp(
-		grpcServer,
-		grpcGateway,
-		cronServer,
-		logger,
-	)
-	if err := server.Run(); err != nil {
-		logger.WithError(err).Fatal("Failed to start server")
-	}
-}
-
-func NewConsulClient(address string) (*api.Client, error) {
-	config := api.DefaultConfig()
-	config.Address = address
-	return api.NewClient(config)
-}
-
-func RegisterService(client *api.Client, serviceName, serviceID, address, serviceType string) error {
-	registration := &api.AgentServiceRegistration{
-		ID:      serviceID,
-		Name:    serviceName,
-		Address: address,
-		Port:    getPortFromAddress(address),
-		Tags:    []string{serviceType},
-		Check: &api.AgentServiceCheck{
-			HTTP:     fmt.Sprintf("http://%s/healthz", address),
-			Interval: "10s",
-			Timeout:  "5s",
-		},
-	}
-
-	if serviceType == "grpc" {
-		registration.Check.GRPC = fmt.Sprintf("%s/grpc.health.v1.Health/Check", address)
-		registration.Check.Interval = "15s"
-	}
-
-	return client.Agent().ServiceRegister(registration)
-}
-
-func DeregisterService(client *api.Client, serviceID string) error {
-	return client.Agent().ServiceDeregister(serviceID)
-}
-
-func getPortFromAddress(address string) int {
-	var port int
-	_, err := fmt.Sscanf(address, ":%d", &port)
+	err = consul.RegisterService(client, serviceName, serviceID, serviceAddress, serviceType, serviceTagsSlice)
 	if err != nil {
-
-		return 0
+		logger.WithError(err).Fatal("Failed to register service with Consul")
 	}
-	return port
+	logger.Infof("Service %s registered with Consul", serviceName)
+
+	go func() {
+		<-time.After(time.Second * 1)
+		if err := consul.DeregisterService(client, serviceID); err != nil {
+			logger.WithError(err).Error("Failed to deregister service")
+		} else {
+			logger.Infof("Service %s deregistered from Consul", serviceName)
+		}
+	}()
+
+	return client
 }
+func initApp(logger *logrus.Logger, dbPool *pgxpool.Pool, consulClient *api.Client) *app.App {
+	grpcAddress := env.GetEnv("LMS_GRPC_SERVER_ADDRESS", ":9001")
+	jwtSecret := env.GetEnv("LMS_JWT_SECRET", "default-secret-key-at-least-32-characters-long")
+	jwtIssuer := env.GetEnv("LMS_JWT_ISSUER", "lms-system")
+	jwtAudience := env.GetEnv("LMS_JWT_AUDIENCE", "lms-client")
 
-func DatabaseConn(logger *logrus.Logger) *pgxpool.Pool {
-	//user := env.GetEnv("LMS_DB_USER", "")
-	//password := env.GetEnv("LMS_DB_PASSWORD", "")
-	//dbName := env.GetEnv("LMS_DB_NAME", "")
-	//host := env.GetEnv("LMS_DB_HOST", "")
-	//port := env.GetEnv("LMS_DB_PORT", "")
-	maxConns := env.GetEnvAsInt("LMS_DB_MAX_CONNS", 10)
-	minConns := env.GetEnvAsInt("LMS_DB_MIN_CONNS", 2)
-	maxConnLifetime := env.GetEnvAsDuration("LMS_DB_MAX_CONN_LIFETIME", time.Hour)
-	maxConnIdleTime := env.GetEnvAsDuration("LMS_DB_MAX_CONN_IDLE_TIME", 30*time.Minute)
-
-	/*dbUrl := fmt.Sprintf("postgres://%s:%s@%s:%s/%s", user, password, host, port, dbName)*/
 	dbUrl := "postgresql://neondb_owner:npg_MEB4CYJS7TKh@ep-withered-salad-a27oo7gn-pooler.eu-central-1.aws.neon.tech/neondb?sslmode=require&channel_binding=require"
-	config, err := pgxpool.ParseConfig(dbUrl)
-	if err != nil {
-		logger.WithError(err).Fatal("Failed to parse database configuration")
-	}
 
-	config.MaxConns = int32(maxConns)
-	config.MinConns = int32(minConns)
-	config.MaxConnLifetime = maxConnLifetime
-	config.MaxConnIdleTime = maxConnIdleTime
-	config.HealthCheckPeriod = 1 * time.Minute
+	backupHandler := handler.NewCornHandler(logger)
+	cronServer := cornServer.NewServer(logger, env.GetEnv("LMS_REDIS_ADDRESS", ""), backupHandler, dbUrl)
 
-	config.ConnConfig.DefaultQueryExecMode = pgx.QueryExecModeCacheStatement
+	store := db.NewStore(logger, dbPool)
+	grpcSrv := rpc.NewServer(store, logger, jwtSecret, jwtIssuer, jwtAudience, grpcAddress)
+	grpcGateway := gateway.NewGateway(logger, grpcAddress, ":8086")
 
-	config.BeforeAcquire = func(ctx context.Context, conn *pgx.Conn) bool {
-		err := conn.Ping(ctx)
-		if err != nil {
-			logger.WithError(err).Warn("Releasing unhealthy connection")
-		}
-		return err == nil
-	}
-
-	var connPool *pgxpool.Pool
-	maxRetries := 5
-	retryDelay := 5 * time.Second
-
-	for i := 0; i < maxRetries; i++ {
-		connPool, err = pgxpool.NewWithConfig(context.Background(), config)
-		if err == nil {
-			break
-		}
-
-		logger.WithError(err).Warnf("Failed to connect to database (attempt %d/%d)", i+1, maxRetries)
-		if i < maxRetries-1 {
-			time.Sleep(retryDelay)
-		}
-	}
-
-	if err != nil {
-		logger.WithError(err).Fatal("Failed to connect to database after retries")
-	}
-
-	if err := connPool.Ping(context.Background()); err != nil {
-		logger.WithError(err).Fatal("Failed to ping database")
-	}
-
-	logger.Info("Successfully connected to database")
-	return connPool
+	return app.NewApp(grpcSrv, grpcGateway, cronServer, logger)
 }
