--- backend/lms-main-system/internal/mapper/dbtomodel.go
@@ -1,1 +1,0 @@
-package mapper
