+++ backend/lms-main-system/proto/authentication/service.authentication.proto
@@ -51,7 +51,6 @@ option (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_swagger) = {
 };
 
 service AuthenticationService {
-  // Register a new user account
   rpc Register(RegisterRequest) returns (RegisterResponse) {
     option (google.api.http) = {
       post: "/lms/v1/auth/register"
